+++ packages/network-controller/src/create-network-client.ts
@@ -1,5 +1,6 @@
 import type { InfuraNetworkType } from '@metamask/controller-utils';
 import { ChainId } from '@metamask/controller-utils';
+import type { PollingBlockTrackerOptions } from '@metamask/eth-block-tracker';
 import { PollingBlockTracker } from '@metamask/eth-block-tracker';
 import { createInfuraMiddleware } from '@metamask/eth-json-rpc-infura';
 import {
@@ -55,36 +56,47 @@ export type NetworkClient = {
  * @param args.configuration - The network configuration.
  * @param args.getRpcServiceOptions - Factory for constructing RPC service
  * options. See {@link NetworkControllerOptions.getRpcServiceOptions}.
+ * @param args.getBlockTrackerOptions - Factory for constructing block tracker
+ * options. See {@link NetworkControllerOptions.getBlockTrackerOptions}.
  * @param args.messenger - The network controller messenger.
- * See {@link NetworkControllerOptions.getRpcServiceOptions}.
+ * @param args.isRpcFailoverEnabled - Whether or not requests sent to the
+ * primary RPC endpoint for this network should be automatically diverted to
+ * provided failover endpoints if the primary is unavailable. This effectively
+ * causes the `failoverRpcUrls` property of the network client configuration
+ * to be honored or ignored.
  * @returns The network client.
  */
 export function createNetworkClient({
   configuration,
   getRpcServiceOptions,
+  getBlockTrackerOptions,
   messenger,
+  isRpcFailoverEnabled,
 }: {
   configuration: NetworkClientConfiguration;
   getRpcServiceOptions: (
     rpcEndpointUrl: string,
   ) => Omit<RpcServiceOptions, 'failoverService' | 'endpointUrl'>;
+  getBlockTrackerOptions: (
+    rpcEndpointUrl: string,
+  ) => Omit<PollingBlockTrackerOptions, 'provider'>;
   messenger: NetworkControllerMessenger;
+  isRpcFailoverEnabled: boolean;
 }): NetworkClient {
   const primaryEndpointUrl =
     configuration.type === NetworkClientType.Infura
       ? `https://${configuration.network}.infura.io/v3/${configuration.infuraProjectId}`
       : configuration.rpcUrl;
-  const availableEndpointUrls = [
-    primaryEndpointUrl,
-    ...(configuration.failoverRpcUrls ?? []),
-  ];
-  const rpcService = new RpcServiceChain(
+  const availableEndpointUrls = isRpcFailoverEnabled
+    ? [primaryEndpointUrl, ...(configuration.failoverRpcUrls ?? [])]
+    : [primaryEndpointUrl];
+  const rpcServiceChain = new RpcServiceChain(
     availableEndpointUrls.map((endpointUrl) => ({
       ...getRpcServiceOptions(endpointUrl),
       endpointUrl,
     })),
   );
-  rpcService.onBreak(({ endpointUrl, failoverEndpointUrl, ...rest }) => {
+  rpcServiceChain.onBreak(({ endpointUrl, failoverEndpointUrl, ...rest }) => {
     let error: unknown;
     if ('error' in rest) {
       error = rest.error;
@@ -99,13 +111,13 @@ export function createNetworkClient({
       error,
     });
   });
-  rpcService.onDegraded(({ endpointUrl }) => {
+  rpcServiceChain.onDegraded(({ endpointUrl }) => {
     messenger.publish('NetworkController:rpcEndpointDegraded', {
       chainId: configuration.chainId,
       endpointUrl,
     });
   });
-  rpcService.onRetry(({ endpointUrl, attempt }) => {
+  rpcServiceChain.onRetry(({ endpointUrl, attempt }) => {
     messenger.publish('NetworkController:rpcEndpointRequestRetried', {
       endpointUrl,
       attempt,
@@ -115,21 +127,19 @@ export function createNetworkClient({
   const rpcApiMiddleware =
     configuration.type === NetworkClientType.Infura
       ? createInfuraMiddleware({
-          rpcService,
+          rpcService: rpcServiceChain,
           options: {
             source: 'metamask',
           },
         })
-      : createFetchMiddleware({ rpcService });
+      : createFetchMiddleware({ rpcService: rpcServiceChain });
 
   const rpcProvider = providerFromMiddleware(rpcApiMiddleware);
 
-  const blockTrackerOpts =
-    process.env.IN_TEST && configuration.type === NetworkClientType.Custom
-      ? { pollingInterval: SECOND }
-      : {};
-  const blockTracker = new PollingBlockTracker({
-    ...blockTrackerOpts,
+  const blockTracker = createBlockTracker({
+    networkClientType: configuration.type,
+    endpointUrl: primaryEndpointUrl,
+    getOptions: getBlockTrackerOptions,
     provider: rpcProvider,
   });
 
@@ -160,6 +170,43 @@ export function createNetworkClient({
   };
 
   return { configuration, provider, blockTracker, destroy };
+}
+
+/**
+ * Create the block tracker for the network.
+ *
+ * @param args - The arguments.
+ * @param args.networkClientType - The type of the network client ("infura" or
+ * "custom").
+ * @param args.endpointUrl - The URL of the endpoint.
+ * @param args.getOptions - Factory for the block tracker options.
+ * @param args.provider - The EIP-1193 provider for the network's JSON-RPC
+ * middleware stack.
+ * @returns The created block tracker.
+ */
+function createBlockTracker({
+  networkClientType,
+  endpointUrl,
+  getOptions,
+  provider,
+}: {
+  networkClientType: NetworkClientType;
+  endpointUrl: string;
+  getOptions: (
+    rpcEndpointUrl: string,
+  ) => Omit<PollingBlockTrackerOptions, 'provider'>;
+  provider: SafeEventEmitterProvider;
+}) {
+  const testOptions =
+    process.env.IN_TEST && networkClientType === NetworkClientType.Custom
+      ? { pollingInterval: SECOND }
+      : {};
+
+  return new PollingBlockTracker({
+    ...testOptions,
+    ...getOptions(endpointUrl),
+    provider,
+  });
 }
 
 /**